Fixed the 2 bugs that Lee detected while doing the Impact Flash animation:
1)  Progressively fewer instanced solids were having their stp
re-referenced because d_uses was not being zeroed by the "clean" command
in the animation script at the start of each new frame.

2) rt_clean() wasn't calling rt_clean_resource() for the per-CPU
resource structures (including the list of bu_bitv structures used for
re_solid_bitv) because of using the wrong macro to see if there were any
to do.  While fixing it I changed tactics so that a semaphored critical
section is no longer needed for this.

svn:revision:17662
cvs:account:mike
cvs:branch:trunk