Converted clients[] array into pointers to PKG connections, rather than container structs with a duplicate of the fd.
Fixes the problem that Keith reported, where fbserv as started from
inetd.conf would connect with client, but not take any data.
In that case, fd==0, which was used as the marker for an empty slot
in the clients[] array.  Thus, the select() input indication would
always be skipped over, and no PKGs would be read off the network.

svn:revision:19037
cvs:account:mike
cvs:branch:trunk