Add a class which outputs plot files to visualize various stages of brep boolean evaluation and use it in the boolean/intersection code.
This is the first piece of the implementation for a debug command which visualizes various stages of a boolean evaluation. The class implementation duplicates a few librt functions to libbrep, and uses them to write out a slew of plot files. This was the path of least developer effort for persistently storing the information to be visualized. To make the command releasable, we'd probably alter it to stash whatever information is needed to create the plots in the gedp, create the plots via librt after evalution has completed, and then store them in persistent hidden binary objects.

svn:revision:61367
svn:branch:brep-debug
svn:account:n_reed