a number of bugs have crept into remrt/rtsrv, ostensibly due to lack of testing.  here we fix one where bu_argv_from_string() is asked to parse a string into an argv, but is off-by-one in the max argv count (passing zero, resulting in no gettrees happening).  the fix doesn't pretend to know what bu_argv_from_string()'s delimiting criteria are, so it sets the max based on string length instead of spaces assuming more than half of them could be delimiters.
due to the way the off-by-one was processed, there happens to be a
user workaround for this bug.  if a file.g provided to remrt is copied
to contain a space (i.e., "file.g file.g") and specified with it, then
rtsrv will parse and dirbuild "file.g".  not pretty, but works.

svn:revision:70900
svn:branch:trunk
svn:account:brlcad