One of the major hold-ups for using Xft/Freetype by default in Tk has been fontconfig - when we mix our bundled freetype (and libpng and zlib) with a system fontconfig it's generally no-go.
A bundled fontconfig has been a daunting challenge to date - its build is
multi-step, using Perl to generate headers and autotools for the main build
flow.  Moreover, at compile time directories are hardcoded where fontconfig
expects to look for fonts and cache entries. Both of these need to be addressed
somehow for a 'bundled' fontconfig to work.

This commit is the beginnings of addressing the problem, starting with the
build process.  Rather than attempt to duplicate the full build process, we'll
take a stripped down, minimalist approach of pulling in any generated headers
we can unless/until there is some clear utility in implementing all the build
steps in a more self contained manner.  Building is untested on Windows (it may
not be needed there, I'll have to check Tk to see).

The second part of the embedding - looking for fonts and processing them -
hasn't been looked at yet - right now if you enable this mged will segfault
with the message "Fontconfig error: Cannot load default config file".  It will
need some design thought, since what we actually want for this is a relative
path runtime lookup in a share/fonts directory in BRL-CAD's runtime hierarchy.
It's not clear yet what the best way to achieve that is.

svn:revision:75266
svn:branch:trunk
svn:account:starseeker