Adjust the layout of the overlap checker's launch menu to avoid a flickering update loop that would occur when hovering over the "Use Last File" button.
A long pathname displayed in the hint text when hovering over the button was causing the window to resize, which would move the button out from underneath the cursor, reverting the hint text to a shorter string, moving the button back under the cursor, triggering the longer hint text, ad infinitum. Now the buttons stay in place when the window auto-resizes.

svn:revision:75477
svn:branch:trunk
svn:account:n_reed