fix idents crash when running from read-only dir
if the output file could not be created, an attempt to free an
unallocate and uninitialized ptbl would cause a bu_bomb.  now its on
the stack, so nothing to free, and error messages are updated to be a
bit more helpful.

svn:revision:76627
svn:branch:trunk
svn:account:brlcad