add support for empty cells in formatting.
this is something strtok() doesn't intrinsically support, so we augment
behavior by looking for sequences of nulls or separators within the
strtok-modified buffer.  strsep() does support empties, but it's not
portably guaranteed.  the reason for implementing this is two-fold: 
first, it lets authors skip columns while printing and second, it
avoids a bug in libfort (reported) that precludes indexing directly
into empty cells.  this was a bit nasty to figure out and not the
best coding to say the least, but not seeing a good way to reduce the
complexity just yet.  it does seem to work well (at least on mac's bsd
strtok), but does warrant portability / unit testing.

svn:revision:77827
svn:branch:trunk
svn:account:brlcad