Make a stab at a portable librt timer.
This code attempts to replace the various librt timer* files with a
single implementation in C++.  Insofar as we can, use standard C/C++
functions to retrieve values.

Windows presents a complication in that clock() does not return CPU
time.  Per their documentation and examples, try to use the only
available option on that platform to get CPU time.

For elapsed time (I think this is intended to be wall clock time?)
use C++11 chrono support which should be portable across all targeted
platforms.

This needs review - I'm not sure we were returning CPU time from the
rt_get_timer function even on systems with a standards compliant
clock() available previously, so this may be a behavior change even
though it is trying to match the behavior described in rt/timer.h.

svn:revision:77862
svn:branch:trunk
svn:account:starseeker