Revert compatibility changes for upstream stepcode.
After discussions with Sean, it's not clear these methods would be
regarded as best practice and we may simply change upstream to
re-expose the direct pointers.

Also, rather than altering our copy, the promising results with the
extbuild experiments suggest it is time to look at getting back in
sync with upstream.  However, rather than doing this in extbuild
and conflating it with those changes, the decision is to do the
stepcode sync in a separate branch. This will allow for easier
testing and change review.

svn:revision:77950
svn:branch:trunk
svn:account:starseeker