Experiment with an option to set working dir.
Trying to get the MSI installer from WiX to change the working
directory for shortcuts has proven challenging.  An alternative
is to allow for the possibility of defining a shortcut that tells
the application itself to change the working directory - untested
as yet.

svn:revision:78040
svn:branch:trunk
svn:account:starseeker