Note work on librt timer code
This is borderline - certain timing outputs may change now
that we are making more of an effort to use the CPU timers
when the API says we are, but it's not clear that there will
be significant user visible implications.

If timing answer change that will be potentially significant,
so going ahead and noting in NEWS to provide some visibility.

svn:revision:78196
svn:branch:RELEASE
svn:account:starseeker